Moved kafka to new package in user-service
